% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova2dt.R
\name{anova2dt}
\alias{anova2dt}
\title{ANOVA results table to `data.table`}
\usage{
anova2dt(tbl, sig.thresh = 0.05)
}
\arguments{
\item{tbl}{anova or data.frame; a table with results from `anova`.}

\item{sig.thresh}{numeric; threshold for marking p-values as significant. Default is 0.05.}
}
\description{
Make a nice looking `data.table` from a results table from `anova`. This function requires the packages 'data.table' and 'stringr' to be installed, and does __not__ work with results from `summary`.
}
\examples{
df <- data.frame(X = c(rnorm(50), {rnorm(50) - 0.8}), Y = c(rep("A", 50), rep("B", 50)))
res <- anova(aov(X ~ Y, data = df))
anova2dt(res)
}
\seealso{
\code{\link{anova}}, \code{\link{as.data.table}}
}
