% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_obj.R
\name{as.obj}
\alias{as.obj}
\title{string as an object name}
\usage{
as.obj(x)
}
\arguments{
\item{x}{a string or object containing a string}
}
\description{
Treat a string as an object name
}
\examples{
a <- 1
b <- "a"
as.obj(b)
}
\seealso{
\code{\link{eval}}, \code{\link{parse}}
}
