% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_n.R
\name{cat.n}
\alias{cat.n}
\title{newline cat}
\usage{
cat.n(x)
}
\arguments{
\item{x}{a string or object containing a string}
}
\description{
Automatically set the `sep` argument in `cat` to `"\n"`
}
\examples{
string <- "hello world!"
cat(string)
cat.n(string)
}
\seealso{
\code{\link{cat}}
}
