% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kci.R
\name{cdkci}
\alias{cdkci}
\title{Conditional independence testing for causal discovery in trams}
\usage{
cdkci(Y, E, X, data, coin = RCIT, verbose = FALSE, ...)
}
\arguments{
\item{Y}{Response}

\item{E}{Environments}

\item{X}{Predictors}

\item{data}{Data}

\item{coin}{Conditional Independence Test}

\item{verbose}{Toggle berbose output; defaults to \code{TRUE}}

\item{...}{Further arguments supplied to \code{coin}}
}
\value{
See KCI
}
\description{
Conditional independence testing for causal discovery in trams
}
\examples{
\dontrun{
set.seed(1)
d <- dgp_random_dag(n = 1e3, mod = "colr", dag = random_dag(nanc = 2,
    ndec = 2, panc = 1, pdec = 1, penv = 1))
cdkci("Y", "E", paste0("X", 1:4), data = d, seed = 1)
}


}
