% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfi_fit.R
\name{cfi_fit}
\alias{cfi_fit}
\title{CFI FIT}
\usage{
cfi_fit(x)
}
\arguments{
\item{x:}{a numeric value indicating the CFI value}
}
\value{
a string describing the cfi value
}
\description{
This function allows you to check the model fit through RMSEA.
}
\examples{
cfi_fit(0.91)
cfi_fit(0.96)
cfi_fit(0.88)
}
