% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chance_to_happen.R
\name{chance.to.happen}
\alias{chance.to.happen}
\title{chance to happen}
\usage{
chance.to.happen(prob, events, round.to = 3, value = FALSE)
}
\arguments{
\item{prob}{numeric; the probability (as a fraction or decimal) of the outcome occurring in a single event}

\item{events}{integer; the number of events}

\item{round.to}{integer; the number of decimal places to round the output to}

\item{value}{logical; if TRUE, return the value of the calculated chance or, if FALSE, print the results}
}
\description{
Calculate the probability of an outcome happening given its per-event probability and the number of events
}
\examples{
# Chance to flip 10 heads in a row
chance.to.happen(prob = 0.5, events = 10)
}
