% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeCohortOrder.R
\name{ChangeCohortOrder}
\alias{ChangeCohortOrder}
\title{Change order of the cohorts in a dataset to "BiB", "EDEN", "INMA", "KANC", "RHEA"}
\usage{
ChangeCohortOrder(input_data)
}
\arguments{
\item{input_data}{A dataframe}
}
\value{
A dataframe with new order of the cohorts
}
\description{
Change order of the cohorts in a dataset to "BiB", "EDEN", "INMA", "KANC", "RHEA"
}
