% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_unused_packages.R
\name{check.unused.packages}
\alias{check.unused.packages}
\title{check for unused pacak}
\usage{
check.unused.packages(files, value = FALSE)
}
\arguments{
\item{x}{a string or object containing a string}
}
\description{
Automatically set the `sep` argument in `cat` to a newline
}
\examples{
string <- "hello world!"
cat(string)
cat.n(string)
}
\seealso{
\code{\link{cat}}
}
