% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpeRs_Visualizers.R
\name{cols2numeric}
\alias{cols2numeric}
\title{Implements...}
\usage{
cols2numeric(...)
}
\arguments{
\item{x}{Description}
}
\value{
\code{z} Description
}
\description{
Implements...
}
\details{
\code{cols2numeric} implements...
}
\examples{

# Perform analysis
cols2numeric()

}
