% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datain.R
\name{datain_filter}
\alias{datain_filter}
\title{Filter datain}
\usage{
datain_filter(datain, id_filter = NULL)
}
\arguments{
\item{datain}{data.frame}

\item{id_filter}{icu regex or NULL}
}
\value{
data.table
}
\description{
Filter datain
}

