% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shiny_utils.R
\name{disabledNumericInput}
\alias{disabledNumericInput}
\title{Return disabled numericInput}
\usage{
disabledNumericInput(numericInput)
}
\arguments{
\item{shiny::disabledNumericInput}{}
}
\value{
shiny::disabledNumericInput
}
\description{
Return disabled numericInput
}

