% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretise.R
\name{discretise}
\alias{discretise}
\title{Discretise an ordered variable}
\usage{
discretise(x, w = NULL, bins = 10, type = NULL, q = NULL, ...)
}
\arguments{
\item{x}{\code{numeric} vector to discretise.}

\item{w}{\code{numeric} vector of weights for each element of x. Default =
\code{NULL}, uses equal weights.}

\item{bins}{\code{numeric} number of bins to split x into.}

\item{type}{\code{character} type of bucketing scheme to use. Possible values
are; 'equal_weight', 'equal_width', 'quantiles'.}

\item{q}{\code{numeric} vector of quantiles to use as break points for bins
if \code{type} is \code{"quantiles"}.}

\item{...}{other arguments passed down to cut for the equal width and
quantiles cases.}
}
\value{
\code{ordered}, \code{factor} version of x, bucketed into the specified
number of bins.
}
\description{
Discretise (bucket) an ordered variable into a set number of bins. Possible
binning schemes are equal width bins, equal weight bins and quantiles.
}
\details{
Currently ordered factors not supported. The actual number of returns bins
may be different to that input depending on the distribution of data values.
}
\examples{
set.seed(1)
a = runif(200)
b = rpois(200, 3)
table(b)
sum(b) / 25
bins = discretise(x = a, w = b, bins = 25, type = 'equal_weight')
table(bins)
tapply(b, bins, sum)

}
