% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_html_table.R
\name{display.html.table}
\alias{display.html.table}
\title{display HTML table in an Rmarkdown document}
\usage{
display.html.table(file, width = 800, height = 400)
}
\arguments{
\item{file}{character; an HTML file}

\item{width}{integer; width of the iframe, in pixels; default 800}

\item{height}{integer; height of the iframe, in pixels; default 400}
}
\description{
Embeds a HTML table (from an HTML file) in an iframe within the output document to preserve nice formatting
}
\examples{
mat <- matrix(rnorm(1:50), nrow = 5, ncol = 10)
flextable::save_as_html(flextable::flextable(as.data.frame(mat)), path = "example_table.html")
display.html.table("example_table.html")
}
\seealso{
\code{\link{readLines}}, \code{\link{URLencode}}
}
