% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{extract_genes}
\alias{extract_genes}
\title{Create tidy table with gene ids and optional chdir coefficients}
\usage{
extract_genes(dt)
}
\arguments{
\item{dt}{tbl_dt as returned from preprocess_data}
}
\value{
tbl_dt with 4 columns: id, category, gene, chdir
}
\description{
Create tidy table with gene ids and optional chdir coefficients
}

