% Generated by manual editing
% Please edit documentation in R/helpeRs_HelperHelpers.R
\name{f2n}
\alias{f2n}
\title{Convert factors to numeric}
\usage{
f2n(x)
}
\arguments{\item{x}{Vector to convert.}}
\value{Numeric vector.}
\description{Converts input to character and then to numeric.}
\examples{\dontrun{f2n(factor(c("1","2")))}}
