% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dt.R
\name{format_dt}
\alias{format_dt}
\title{format_dt}
\usage{
format_dt(x)
}
\arguments{
\item{x}{a data object (either a \code{matrix} or a \code{data.frame})}
}
\value{
a data table decorated with export buttons
}
\description{
Turns a data frame or matrix into a DT formatted table
with export buttons for copy, csv or excel.
}
\examples{
format_dt(USArrests)

}
\author{
Denis O'Meally
}
