% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_divisors.R
\name{get.divisors}
\alias{get.divisors}
\title{get divisors}
\usage{
get.divisors(x)
}
\arguments{
\item{x}{integer}
}
\description{
Get the divisors of an integer
}
\examples{
color.pal <- getCB.pal("black")
}
