% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{get_path}
\alias{get_path}
\title{Download file if required and return path to directory}
\usage{
get_path(sigs_path, choice)
}
\arguments{
\item{sigs_path}{character}

\item{choice}{character}
}
\value{
character
}
\description{
Download file if required and return path to directory
}

