% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorblind_pals.R
\name{getCB.pal}
\alias{getCB.pal}
\title{colorblind palettes}
\usage{
getCB.pal(pal = c("grey", "black"))
}
\arguments{
\item{pal}{string or integer; choose either the `"grey"` or `"black"` palette. Can also be `"g"`, `"b"`, `"1"`, or `"2"`, respectively}
}
\description{
Select from two colorblind-friendly palettes (8 colors each) with either gray or black as the first color
}
\examples{
color.pal <- getCB.pal("black")
}
