% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpeRs_Visualizers.R
\name{heatMap}
\alias{heatMap}
\title{Implements...}
\usage{
heatMap(...)
}
\arguments{
\item{x}{Description}
}
\value{
\code{z} Description
}
\description{
Implements...
}
\details{
\code{image2} implements...
}
\examples{

# Perform analysis
heatMap()

}
