% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{package}
\name{helpers}
\alias{helpers}
\alias{helpers-package}
\title{helpers: General data analysis functions for R.}
\description{
This package contains some functions that I commonly use when doing analysis
in R.
}
\details{
This includes functions to discretise ordered variables, summarise variables
in a data.frame (i.e. group by variable and sum other columns) and plot these
variable summaries.
}
