% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsehist.R
\name{hist_search}
\alias{hist_search}
\title{History searching.}
\usage{
hist_search(search_string = "ccle_cn_endo<-", location = ".")
}
\arguments{
\item{search_string="ccle_cn_endo<-"}{String to be searched for.}
}
\description{
Searching for string in a multitude of horrible horrible history files. A fancy grep.
}
\examples{
hist_search(search_string="horrible_var_name")
}
\keyword{history,}
\keyword{search}
