% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpeRs_Visualizers.R
\name{image2}
\alias{image2}
\title{Implements...}
\usage{
image2(...)
}
\arguments{
\item{x}{Description}
}
\value{
\code{z} Description
}
\description{
Implements...
}
\details{
\code{image2} implements...
}
\examples{

# Perform analysis
image2()

}
