% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpeRs_Visualizers.R
\name{MakeHeatMap}
\alias{MakeHeatMap}
\title{Implements...}
\usage{
MakeHeatMap(...)
}
\arguments{
\item{x}{Description}
}
\value{
\code{z} Description
}
\description{
Implements...
}
\details{
\code{image2} implements...
}
\examples{

# Perform analysis
MakeHeatMap()

}
