% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_helpers.R
\name{matrix2tpm}
\alias{matrix2tpm}
\title{matrix2tpm}
\usage{
matrix2tpm(counts, len)
}
\arguments{
\item{len}{a vector of feature lengths, in Kbp}

\item{data_matrix}{raw counts, as a data.frame or matrix}
}
\value{
a matrix of TPMs
}
\description{
calculate TPMs from a matrix and a vector of feature lengths
}
\author{
Mike Love
}
