% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mids2SingleComplete.R
\name{Mids2SingleComplete}
\alias{Mids2SingleComplete}
\title{Create a one (number 20) complete imputed dataset (from mids object)}
\usage{
Mids2SingleComplete(input_data, variable_list, indx = 20)
}
\arguments{
\item{input_data}{A mids object}

\item{variable_list}{Exposure variables to be analysed}

\item{indx}{A scalar from 1 to n of imputed datasets, choosing which dataset will be returned}
}
\value{
A 'random' complete dataset (number 20)
}
\description{
Create a one (number 20) complete imputed dataset (from mids object)
}
