% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min2time.R
\name{min2time}
\alias{min2time}
\title{convert decimal minutes to long form}
\usage{
min2time(m, include.year = FALSE)
}
\arguments{
\item{m}{numeric; minutes in decimal form}

\item{include.year}{logical; whether to include years in the output (TRUE), or just days, hours, minutes, and seconds (FALSE)}
}
\description{
Convert minutes in decimal form to (years), days, hours, minutes, and seconds
}
\examples{
min2time(527.86)
}
