% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_assoc.R
\name{mixed_assoc}
\alias{mixed_assoc}
\title{Mixed pairwise association}
\usage{
mixed_assoc(df, cor_method = "spearman", adjust_cramersv_bias = TRUE)
}
\arguments{
\item{df}{a data.frame}

\item{cor_method}{the method to use, one of \code{c("pearson", "kendall", "spearman")}; (Default: \code{spearman})}

\item{adjust_cramersv_bias}{Use bias corrected Cramer's V as described in \code{\link[rcompanion:cramerV]{rcompanion::cramerV()}}; (Default: \code{TRUE})}
}
\value{
a list with two elements:
\itemize{
\item \code{associations} a data.frame  with the pairwise association between each variable
\item \code{assoc_matrix} a square matrix of association scores between each variable, with dimensions
equal to the number of variables.
}
}
\description{
Calculate a pairwise association between all variables in a data-frame.
In particular nominal vs nominal with Chi-square, numeric vs numeric
with Pearson correlation, and nominal vs numeric with ANOVA.
Adapted from \url{https://stackoverflow.com/a/52557631/590437}
}
