% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{month_char}
\alias{month_char}
\title{Convert month integers to two digit strings}
\usage{
month_char(month)
}
\arguments{
\item{month}{integer vector; allowed values 1 to 12.}
}
\value{
character vector with 0 appended to the start of values less than 10.
}
\description{
Converts month from integer format to two digit string format. 1 to '01'. 11 to '11'.
}
\examples{
month_char(1:12)
}
