% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{month_label}
\alias{month_label}
\title{Convert year and month to month label}
\usage{
month_label(year, month, abb = FALSE, upper_case = FALSE)
}
\arguments{
\item{year}{integer vector}

\item{month}{integer vector contain values from 1 to 12}

\item{abb}{if TRUE month abbreviation is used instead of full name (default FALSE)}

\item{upper_case}{if TRUE month abbreviation or name is upper case}
}
\value{

}
\description{
Converts year and month to 'year-two digit month number-month abbreviation or name
}
\examples{
month_label(2021, 6)
month_label(2021, 6, abb = TRUE)
month_label(2021, 6, upper_case = TRUE)
}
