% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nds.R
\name{nds}
\alias{nds}
\title{roll n dice of side s}
\usage{
nds(n = 1, s = 4, sum = TRUE)
}
\arguments{
\item{n}{integer; number of dice to roll.}

\item{s}{integer; number of sided die to roll.}

\item{sum}{logical; if TRUE, return the sum of the results, if false, return a vector of results equal in length to `n`. Default is TRUE.}
}
\description{
Roll some dice!
}
\examples{
# Roll 4d6!
nds(n = 4, s = 6)
}
\seealso{
\code{\link{sample}}
}
