% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datain.R
\name{prepare_density_plot_input}
\alias{prepare_density_plot_input}
\title{Prepare input for density plot}
\usage{
prepare_density_plot_input(datain)
}
\arguments{
\item{datain}{data.table with genenames in first columns and samples in following}
}
\value{
data.frame
}
\description{
Prepare input for density plot
}

