% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{prepare_enrichr_input}
\alias{prepare_enrichr_input}
\title{Prepare input for Enrichr form}
\usage{
prepare_enrichr_input(chdir_results)
}
\arguments{
\item{chdir_results}{data.frame as returned from prepare_results}
}
\value{
character
}
\description{
Prepare input for Enrichr form
}

