% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Convert do tbl_dt, add names, and clean strings}
\usage{
preprocess_data(dt)
}
\arguments{
\item{dt}{data.table}
}
\value{
tbl_dt
}
\description{
Convert do tbl_dt, add names, and clean strings
}

