% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_tbl.R
\name{prev.tbl}
\alias{prev.tbl}
\title{preview table}
\usage{
prev.tbl(tbl, n = 6)
}
\arguments{
\item{tbl}{an array, matrix, data.frame, data.table, tibble, etc}

\item{n}{integer; the number of rows and columns you want to preview, default 6}
}
\description{
Previews the first n rows and columns from an array, matrix, data.frame...etc. Like `head` but also limits the columns shown, not just the rows.
}
\examples{
mat <- matrix(rnorm(1:10000), nrow = 100, ncol = 100)
prev.tbl(mat)
}
\seealso{
\code{\link{head}}
}
