% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nds.R
\name{"\%d\%"}
\alias{"\%d\%"}
\alias{\%d\%}
\title{roll n dice of side s}
\usage{
n \%d\% s
}
\arguments{
\item{n}{integer; number of dice to roll}

\item{s}{integer; number of sided die to roll}
}
\description{
Roll some dice!
}
\examples{
# Roll 4d6!
4\%d\%6
}
\seealso{
\code{\link{sample}}
}
