% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp-random-dag.R
\name{random_dag}
\alias{random_dag}
\title{Title}
\usage{
random_dag(
  nenv = 2,
  nanc = 3,
  ndec = 3,
  penv = 0.3,
  panc = 0.3,
  pdec = 0.3,
  cfx = .rcfx(nanc, panc, ensure_one = TRUE),
  cfc = .rcfx(ndec, pdec),
  cfb = c(-6, 4),
  cfe = .rcfx(nanc + ndec, penv),
  plot_graph = FALSE,
  BanYdeY = matrix(rnorm(nanc * ndec, sd = sqrt(0.5)), nrow = nanc, ncol = ndec),
  AanYdeY = matrix(sample(0:1, nanc * ndec, replace = TRUE), nrow = nanc, ncol = ndec)
)
}
\arguments{
\item{nenv}{Number of environments}

\item{nanc}{Number of potential ancestors}

\item{ndec}{Number of potential descendants}

\item{penv}{Edge probability for environment}

\item{panc}{Edge probability for ancestral graph}

\item{pdec}{Edge probability for descendental graph}

\item{cfx}{Coefficients from X to Y}

\item{cfc}{Coefficients from Y to children}

\item{cfb}{Coefficients of baseline trafo in case of (log-) linear basis}

\item{cfe}{Coefficients from E to X}

\item{plot_graph}{Toggle, plot DAG}

\item{BanYdeY}{Coefficient matrix for ancestral to descendental variables}

\item{AanYdeY}{Masking matrix for 0 coefs in \code{BanYdeY}}
}
\value{
List containing random DAG and infos for simulation experiments
}
\description{
Title
}
\examples{
random_dag()

}
