% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactoring.R
\name{refactor}
\alias{refactor}
\title{A Refactor Function}
\usage{
refactor(a_factor)
}
\arguments{
\item{a_factor}{Factor variable to be changed.}
}
\description{
Factor levels trimming
}
\examples{
refactor(education_types)
}
\keyword{factors,}
\keyword{levels}
