% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.r
\name{remove_from}
\alias{remove_from}
\title{Removes values from a vector}
\usage{
remove_from(vec, value)
}
\arguments{
\item{vec}{the vector to remove values from}

\item{value}{the value to remove. The value can either be a single value or
a vector of values}
}
\description{
The \code{remove_} functions are convenient wrappers to help remove specific
values from a vector. \code{remove_from} will remove single or multiple
values from a given vector. In addition the \code{remove_null} and
\code{remove_na} functions will remove the corresponding missing values.
}

