% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_dt_NAs.R
\name{replace.dt.NAs}
\alias{replace.dt.NAs}
\title{replace all NAs in a data.table}
\usage{
replace.dt.NAs(dt, replacement = 0)
}
\arguments{
\item{dt}{a data.table}

\item{replacement}{an integer, number, or character. Default 0}
}
\description{
Replaces all NAs in an object of class data.table with 0 (or another replacement)
}
\examples{
require(data.table)
DT <- data.table(X = c(1:6, NA, 8:10), Y = letters[1:10])
DT.noNA <- replace.dt.NAs(DT)
DT
DT.noNA
}
\seealso{
\code{\link{data.matrix}}, \code{\link{data.table::data.table}}
}
