% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right_triangle.R
\name{right.triangle}
\alias{right.triangle}
\title{right triangle}
\usage{
right.triangle(a = NULL, b = NULL, c = NULL, round.to = 3, value = FALSE)
}
\arguments{
\item{a}{numeric; length of side A}

\item{b}{numeric; length of side B}

\item{c}{numeric; length of side C, the hypotenuse}

\item{round.to}{integer; the number of decimals to round the printed output to}

\item{value}{logical; TRUE: return the value, FALSE: just print the value}
}
\description{
Given the lenghts of two sides of a right triangle, returns the third (via the Pythagorean theorem)
}
\examples{
mat <- matrix(rnorm(1:10000), nrow = 100, ncol = 100)
prev.tbl(mat)

}
