% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_helpers.R
\name{RLEplot_mod}
\alias{RLEplot_mod}
\title{RLEplot_mod}
\usage{
RLEplot_mod(data_matrix, sample_sheet, title = NULL, caption = NULL)
}
\arguments{
\item{data_matrix}{log2 transformed counts, as a data.frame or matrix}

\item{sample_sheet}{a sample_sheet of class data.frame, with a column named batch}

\item{title}{a title for the plot. Leave blank for noe, or a character string}

\item{caption}{caption for the plot. Leave blank for noe, or a character string}
}
\value{
a ggplot object
}
\description{
plot the RLEs for each sample
}
\author{
Denis O'Meally
}
