% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_group.R
\name{sample_group}
\alias{sample_group}
\title{sample from groups of variables}
\usage{
sample_group(df, ..., n)
}
\arguments{
\item{df}{the data.frame you are sampling from}

\item{...}{the variables you would like to sample from}

\item{n}{the number of samples to return}
}
\value{
a data frame containing the samples groupings
}
\description{
\code{sample_group} allows you to sample observations from groupings of
variables. This is useful for when you have multilevel data and would like
to sample from multiple levels at once. This makes exploratory plotting a
little easier.
}
\examples{
sample_group(mtcars, cyl, am)
}

