% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec2time.R
\name{sec2time}
\alias{sec2time}
\title{convert decimal seconds to long form}
\usage{
sec2time(s, include.year = FALSE)
}
\arguments{
\item{s}{numeric; seconds in decimal form}

\item{include.year}{logical; whether to include years in the output (TRUE), or just days, hours, minutes, and seconds (FALSE)}
}
\description{
Convert seconds in decimal form to (years), days, hours, minutes, and seconds
}
\examples{
sec2time(5027.32)
}
