% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_packages.R
\name{source.packages}
\alias{source.packages}
\title{check if all packages have been loaded, and if not, source the input file}
\usage{
source.packages(file)
}
\arguments{
\item{file}{character; path and name of a file that load packages with `library` or `require`}
}
\description{
For use with a script specifically for loading all necessary packages. This functions checks the session info to see if all packages have been loaded. If not, it sources the file.
}
\seealso{
\code{\link{sessionInfo}}
}
