% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/paea.R
\name{split_chdirresults}
\alias{split_chdirresults}
\title{Split chdir vector into negative and postive (plus zero) subvectors}
\usage{
split_chdirresults(chdirresults)
}
\arguments{
\item{chdirresults}{see GeoDE::PAEAAnalysis}
}
\value{
list with up and down fields
}
\description{
Split chdir vector into negative and postive (plus zero) subvectors
}

