% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/paea.R
\name{split_gmtfile}
\alias{split_gmtfile}
\title{Split gmt file into up and down based on _downregulated suffix}
\usage{
split_gmtfile(gmtfile)
}
\arguments{
\item{gmtfile}{see GeoDE::chdirAnalysis}
}
\value{
list with up and down sets
}
\description{
Split gmt file into up and down based on _downregulated suffix
}

