% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stats.R
\name{submissions_per_organism}
\alias{submissions_per_organism}
\title{Count number of submissions per organism}
\usage{
submissions_per_organism(dataset)
}
\arguments{
\item{list}{of dataframes as returned from preprocess}
}
\value{
data.frame with organism and n
}
\description{
Count number of submissions per organism
}

