% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stats.R
\name{submissions_per_user}
\alias{submissions_per_user}
\title{Count number of submissions per user}
\usage{
submissions_per_user(dataset)
}
\arguments{
\item{list}{of dataframes as returned from preprocess}
}
\value{
data.frame with curator and n
}
\description{
Count number of submissions per user
}

