% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_col_dplyr}
\alias{summarise_col_dplyr}
\title{Summarise a column of a data.frame using dplyr}
\usage{
summarise_col_dplyr(df, col, observed, predictions, weights)
}
\arguments{
\item{df}{\code{data.frame} containing variables of interest.}

\item{col}{\code{character} name of column to summarise by.}

\item{observed}{\code{character} name of observed column to summarise by col.}

\item{predictions}{\code{character} name of predictions column to summarise
by col, can be a vector of multiple columns.}

\item{weights}{\code{character} name of weights column to summarise by col.}

\item{ordered_bins}{\code{numeic} number of bins to discretise col into if it
is ordered. Default = \code{20}.}

\item{ordered_binning}{\code{character} type of discretisation to use if col
is ordered. Default = \code{"equal_width"}.}
}
\value{
\code{tibble} containing weights, observed and predictions summarised by col.
}
\description{
Take a particular variable (x) in a data.frame and summarise other variables
of interest by this variable (x). Useful if you want to summarise predictions
and observed values from a model, by certain explanatory variables.
}
\examples{
set.seed(1)
data <- data.frame(a = c(runif(1000), rep(NA, 10)),
                   b = rnorm(1010),
                   c = rpois(1010, 3),
                   d = rnorm(1010),
                   e = runif(1010),
                   f = factor(sample(1010)),
                   g = as.character(sample(5, size = 1010, replace = T)))

summarise_col_dplyr(df = data,
                   col = "g",
                   observed = "b",
                   predictions = c("e","d"),
                   weights = "c")

}
