% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_columns}
\alias{summarise_columns}
\title{Summarise multiple columns of a data.frame}
\usage{
summarise_columns(df, cols = NULL, observed = NULL,
  predictions = NULL, weights = NULL)
}
\arguments{
\item{df}{\code{data.frame} containing variables of interest.}

\item{cols}{\code{character} vector of columns to summarise by.}

\item{observed}{\code{character} name of observed column to summarise by col.}

\item{predictions}{\code{character} name of predictions column to summarise
by col, can be a vector of multiple columns.}

\item{weights}{\code{character} name of weights column to summarise by col.
Default = \code{NULL} uses equal weights.}
}
\value{
named \code{list} containing a \code{tibble} for each column specified in
\code{cols}. Each \code{tibble} contains weights, observed and predictions
(if specified) summarised by that particular column.
}
\description{
Take a serveral variables (x) in a data.frame and summarise other variables
of interest by these variables (x). Useful if you want to summarise
predictions and observed values from a model, by certain explanatory
variables.
}
\examples{
set.seed(1)
data <- data.frame(a = c(runif(1000), rep(NA, 10)),
                   b = rnorm(1010),
                   c = rpois(1010, 3),
                   d = rnorm(1010),
                   e = runif(1010),
                   f = factor(sample(1010)),
                   g = as.character(sample(5, size = 1010, replace = T)))

summarise_column(df = data,
                 col = c("a", "c", "f", "g"),
                 observed = "b",
                 predictions = c("e","d"))

}
