% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_df_values}
\alias{summarise_df_values}
\title{Summary of data.frame values.}
\usage{
summarise_df_values(data, cols = NULL, max_levels = 100,
  print_levels = 5, csv_out = NULL)
}
\arguments{
\item{data}{\code{data.frame} containing columns to summarise.}

\item{cols}{\code{character} names of columns in data to summarise. Default =
\code{NULL} summarises all columns.}

\item{max_levels}{\code{numeric} maximum number of data values to report per
variable.}

\item{print_levels}{\code{numeric} for variables with more levels than
max_levels how many levels should be reported instead? Default value = 5
means than the first, middle and last 5 data values are output for these
variables.}

\item{csv_out}{\code{character} csv file to output results table to. Default
= \code{NULL} means no table is output.}
}
\value{
\code{data.frame} containing variable summaries and optionally csv
version too.
}
\description{
For each requested variable return results from table() in a single
\code{data.frame}. If a given variable has more unique data values than
the cut off (\code{max_levels}), then the first, middle and last
\code{print_levels} will be returned instead. Results are returned in a
single \code{data.frame}.
}
\examples{
set.seed(1)
df <- data.frame(a = rnorm(1000),
                 b = runif(1000),
                 c = sample(5),
                 d = as.factor(sample(5)))

data_levels_summary(data = df,
                    cols = NULL,
                    max_levels = 20,
                    print_levels = 5,
                    csv_out = NULL)
}
