% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_importation.R
\name{tbl_dbGetQuery}
\alias{tbl_dbGetQuery}
\title{Query data from database as a tibble}
\usage{
tbl_dbGetQuery(conn, statement, all_upper = F)
}
\arguments{
\item{conn}{DBI database connection}

\item{statement}{database query (select statement)}

\item{all_upper}{(default FALSE) if TRUE converts all field names to upper case.}
}
\value{
tibble
}
\description{
Wrapper around DBI::dbGetQuery that imports query results as tibbles instead of data.frames.
Also has the option to convert all field names to upper case.
}
