% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_importation.R
\name{tbl_dbReadTable}
\alias{tbl_dbReadTable}
\title{Read table from database as a tibble}
\usage{
tbl_dbReadTable(conn, name, all_upper = F)
}
\arguments{
\item{conn}{DBI database connection}

\item{name}{name of table to import}

\item{all_upper}{(default FALSE) if TRUE converts all field names to upper case.}
}
\value{
tibble
}
\description{
Wrapper around DBI::dbReadTable that imports database tables as tibbles instead of data.frames.
Also has the option to convert all field names to upper case.
}
