% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_importation.R
\name{tbl_fread}
\alias{tbl_fread}
\title{Download flat text file to data.frame}
\usage{
tbl_fread(...)
}
\arguments{
\item{...}{see ?data.table::fread or args(data.table::fread)}
}
\value{

}
\description{
Download flat text file to data.frame
}
