% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unavailable_packages.R
\name{unavailable.packages}
\alias{unavailable.packages}
\title{list (and install) unavailable packages}
\usage{
unavailable.packages(file, auto.install = TRUE, use.bioconductor = TRUE)
}
\arguments{
\item{file}{character; the name of a file that loads packages}

\item{auto.install}{logical; TRUE: will attempt to install all missing packages, FALSE: will ask whether to install all, one, or no missing packages. default TRUE}

\item{use.bioconductor}{logical; TRUE: use `bioconductor` (i.e., function `BiocManager::install`) to install the missing packages, FALSE: use base `install.packages` function. default TRUE}
}
\description{
Finds packages loaded (via `library` or `require`) by a script and determines which ones are not installed. Installs missing packages by default.
}
\seealso{
\code{\link{BiocManager::install}}, \code{\link{install.packages}}
}
