% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpeRs_HelperHelpers.R
\name{vcovCluster}
\alias{vcovCluster}
\title{Cluster-robust covariance matrix estimator}
\usage{
vcovCluster(fm, clvar)
}
\arguments{
\item{fm}{A fitted model object (typically from \code{lm()} or \code{glm()}).
For \code{polr} models from \pkg{MASS}, predictions are handled specially.}

\item{clvar}{Character string giving the name of the clustering variable.
This variable must exist in the data used to fit \code{fm}.}
}
\value{
A K x K covariance matrix where K is the number of model coefficients.
}
\description{
Computes a clustered sandwich covariance matrix for a fitted model using the
method of Arellano (1987). This provides standard errors that are robust to
arbitrary within-cluster correlation.
}
\details{
The function applies the degrees-of-freedom correction
\eqn{(M/(M-1)) \times ((N-1)/(N-K))}{(M/(M-1)) * ((N-1)/(N-K))} where M is
the number of clusters, N is the number of observations, and K is the number
of parameters.
}
\examples{
\dontrun{
data(mtcars)
fit <- lm(mpg ~ wt + hp, data = mtcars)

# Cluster by number of cylinders
V_clust <- vcovCluster(fit, "cyl")

# Use with coeftest for clustered standard errors
library(lmtest)
coeftest(fit, vcov. = V_clust)
}

}
\references{
Arellano, M. (1987). Computing Robust Standard Errors for Within-Groups
Estimators. \emph{Oxford Bulletin of Economics and Statistics}, 49(4), 431-434.
}
\seealso{
\code{\link{GetTableEntry}} which uses this function for clustered
  standard errors, \code{\link[sandwich]{vcovHC}} for heteroskedasticity-robust
  covariance without clustering
}
