% Generated by manual editing
% Please edit documentation in R/helpeRs_HelperHelpers.R
\name{WidenMargins}
\alias{WidenMargins}
\title{Wrap a LaTeX table in an adjustwidth environment}
\usage{
WidenMargins(x)
}
\arguments{\item{x}{Character vector containing the LaTeX table.}}
\value{Modified character vector.}
\description{Adds an \code{adjustwidth} environment around a table to slightly widen the margins.}
\examples{\dontrun{WidenMargins(readLines("table.tex"))}}
