% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_installed_packages.R
\name{write.installed.packages}
\alias{write.installed.packages}
\title{write installed packages to file}
\usage{
write.installed.packages(
  name = "installed_packages",
  path = "./",
  filetype = c("txt", "rds"),
  include.date = TRUE
)
}
\arguments{
\item{name}{character; the name of the output file, NO EXTENSION. Default "installed_packages"}

\item{path}{character; the name of the output directory. Default "./"}

\item{filetype}{character; the type of file to write "txt" for a plaintext file or "rds" for an R binary object file. Default "txt"}

\item{include.date}{logical; whether to include the current date in the output file name. Default TRUE}
}
\description{
Lists and writes all installed packages to a file.
}
\examples{
write.installed.packages()
}
\seealso{
\code{\link{installed.packages}}, \code{\link{saveRDS}}, \code{\link{writeLines}}
}
