% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_helpers.R
\name{write_tpm_matrix}
\alias{write_tpm_matrix}
\title{Write an expression matrix to file}
\usage{
write_tpm_matrix(
  summarised_experiment,
  path = "inst/extdata/",
  drop = NULL,
  tpm = 1,
  samples = 5
)
}
\arguments{
\item{summarised_experiment}{a \code{SummarisedExperiment} object}

\item{path}{the existing folder to write to, Default is \verb{inst/extdata/}}

\item{drop}{samples (columns) to remove}

\item{tpm}{minimum TPM for a gene to be kept, Default: 1}

\item{samples}{minimum number of samples a gene must be present in to be kept, Default: 5}
}
\value{
path to the CSV file
}
\description{
Write an TPM expression matrix to a CSV file in \code{extdata}
}
\details{
This function writes an expression matrix to a CSV file in \code{extdata},
filtering out genes with fewer than \code{tpm} reads in at least \code{samples} samples.
}
\examples{
\dontrun{
if (interactive()) {
    write_tpm_matrix(cml_mrna_GRCm38_HLT, c("F545.7", "X490.17"), tpm = 1, samples = 5)
}
}
}
\seealso{
\code{\link[SummarizedExperiment]{SummarizedExperiment-class}}
\code{\link[utils]{write.table}}
}
