% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpeRs_HelperHelpers.R
\name{fixZeroEndings}
\alias{fixZeroEndings}
\title{Implements...}
\usage{
fixZeroEndings(...)
}
\arguments{
\item{x}{Description}
}
\value{
\code{z} Description
}
\description{
Implements...
}
\details{
\code{fixZeroEndings} implements...
}
\examples{

# Perform analysis
fixZeroEndings <- getTableEntry()

print( fixZeroEndings )

}
