% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_helpers.R
\name{RLEplot_mod}
\alias{RLEplot_mod}
\title{RLEplot_mod}
\usage{
RLEplot_mod(
  data_matrix,
  sample_sheet,
  col_by = group,
  title = NULL,
  caption = NULL
)
}
\arguments{
\item{data_matrix}{log2 transformed counts, as a \code{data.frame} or \code{matrix}}

\item{sample_sheet}{a sample_sheet object that describes the sample metadata}

\item{col_by}{a column name from \code{sample_sheet} to colour the plot by}

\item{title}{a title for the plot. Leave blank for none, or a character string}

\item{caption}{caption for the plot. Leave blank for none, or a character string}
}
\value{
a ggplot object
}
\description{
plot the relative log expresion for each sample. This plot is
useful for comparing normalization procedures between groups.
}
\examples{

m <- -log2(replicate(10, rgamma(3e4, 0.1))) \%>\%
    magrittr::set_colnames(letters[1:10])
sample_sheet <- data.frame(
    sample_id = letters[1:10],
    group = c(rep("A", 4), rep("B", 6))
)

RLEplot_mod(
    m,
    sample_sheet,
    col_by = "group",
    title = "Relative log expression",
    caption = "Simulated data set of 10 samples in 5 groups drawn from 30000 genes"
)
}
\author{
Denis O'Meally
}
