% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsea_fit.R
\name{rmsea_fit}
\alias{rmsea_fit}
\title{RMSEA FIT}
\usage{
rmsea_fit(x)
}
\arguments{
\item{x}{a numeric value indicating the RMSEA value}
}
\description{
This function allows you to check the model fit through RMSEA.
}
\examples{
rmsea_fit(0.04)
}
\keyword{rmsea}
