% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/select_distinct.R
\name{select_distinct}
\alias{select_distinct}
\title{select distinct combinations of multiple variables}
\usage{
select_distinct(df, ...)
}
\arguments{
\item{df}{the data frame you would like to select from}

\item{...}{the variable names you would like to select}
}
\value{
a data frame containing the distinct combination of the variables
that were selected
}
\description{
\code{select_distinct} is just a wrapper around the dplyr verbs
\code{distinct} and \code{select}. \code{select} will trim the variables
you don't want from the data frame and \code{distinct} will return only
the groupings of the variables that exist in the data. This is mainly
useful when you have multiple levels of data that is repeated for each row
in a data frame. Think a data frame containing the information on a school,
class, and student level. You can use \code{select_distinct} to get a data
frame of the classes and school combinations.
}
\examples{
select_distinct(mtcars, cyl, am)
}

