% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperHelpers.R
\name{vcovCluster}
\alias{vcovCluster}
\title{Implements...}
\usage{
getTableEntry(...)
}
\arguments{
\item{x}{Description}
}
\value{
\code{z} Description
}
\description{
Implements...
}
\details{
\code{getTableEntry} implements...
}
\examples{

# Perform analysis
TableEntry <- getTableEntry()

print( TableEntry )

}
