% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpeRs_HelperHelpers.R
\name{vcovCluster}
\alias{vcovCluster}
\title{Implements...}
\usage{
vcovCluster(...)

vcovCluster(...)
}
\arguments{
\item{x}{Description}
}
\value{
\code{z} Description

\code{z} Description
}
\description{
Implements...

Implements...
}
\details{
\code{vcovCluster} implements...

\code{vcovCluster} implements...
}
\examples{

# Perform analysis
TableEntry <- vcovCluster()

print( TableEntry )


# Perform analysis
TableEntry <- vcovCluster()

print( TableEntry )

}
