% Generated by manual editing
% Please edit documentation in R/helpeRs_HelperHelpers.R
\name{vcovCluster}
\alias{vcovCluster}
\title{Cluster-robust covariance matrix estimator}
\usage{
vcovCluster(fm, clvar)
}
\arguments{
  \item{fm}{A fitted model object.}
  \item{clvar}{Name of the clustering variable.}
}
\value{A covariance matrix.}
\description{Computes a clustered sandwich covariance matrix for a fitted model.}
\examples{\dontrun{vcovCluster(lm(y~x,data), "group")}}
